requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
msgpack==1.0.2
requests==2.28.2
signalrcore==0.9.5
urllib3==1.26.15
websocket-client==1.0.0
gym==0.21
pygame==2.1.0
stable-baselines3[extra]==1.8.0
hydra-core==1.3.2

-i https://download.pytorch.org/whl/cu118
torch